changed return type of method getDuration() to int

+++ neo4j-core/src/main/java/at/srfg/graphium/neo4j/model/cache/SegmentCacheEntry.java
@@ -66,7 +66,7 @@ public class SegmentCacheEntry {
 		this.length = length;
 	}
 
-	public float getDuration(boolean directionTow) {
+	public int getDuration(boolean directionTow) {
 		if (directionTow) {
 			return calcDuration(maxSpeedTow);
 		} else {
